--- app/src/main/java/com/stormers/storm/card/network/CardInterface.kt
@@ -1,13 +1,0 @@
-package com.stormers.storm.card.network
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface CardInterface {
-
-    @GET("/card/memo/{user_idx}/{card_idx}")
-    fun responseCardData(
-        @Path("user_idx") user_idx : Int,
-        @Path("card_idx") card_idx : Int) : Call<ResponseCardData>
-}
