+++ app/src/main/java/com/stormers/storm/card/network/RequestCard.kt
@@ -7,4 +7,9 @@ import retrofit2.http.Path
 interface RequestCard {
     @GET("/round/cardList/{project_idx}/{round_idx}")
     fun requestCard(@Path("project_idx") projectIdx : Int, @Path("round_idx") roundIdx: Int) : Call<ResponseCardData>
+
+    @GET("/card/memo/{user_idx}/{card_idx}")
+    fun responseCardData(
+        @Path("user_idx") user_idx : Int,
+        @Path("card_idx") card_idx : Int) : Call<ResponseCardData>
 }
