--- app/src/main/java/com/stormers/storm/network/InterfaceAddProject.kt
@@ -1,12 +1,0 @@
-package com.stormers.storm.network
-
-import com.stormers.storm.project.model.AddProjectModel
-import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface InterfaceAddProject {
-    @POST("/project")
-    fun addProject(@Body body:AddProjectModel) :Call<ResponseAddProject>
-
-}
