--- app/src/main/java/com/stormers/storm/network/InterfaceJoinProjectUsingCode.kt
@@ -1,12 +1,0 @@
-package com.stormers.storm.network
-
-import com.stormers.storm.project.model.JoinProjectUsingCodeModel
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface InterfaceJoinProjectUsingCode{
-    @POST("/project/enter")
-    fun joinProjectUsingCode(@Body body:JoinProjectUsingCodeModel) : Call<ResponseJoinProjectUsingCode>
-}
