+++ app/src/main/java/com/stormers/storm/project/base/BaseProjectWaitingActivity.kt
@@ -7,10 +7,9 @@ import android.view.MotionEvent
 import android.view.inputmethod.InputMethodManager
 import com.stormers.storm.R
 import com.stormers.storm.base.BaseActivity
-import com.stormers.storm.network.InterfaceProjectInfo
+import com.stormers.storm.project.network.RequestProject
 import com.stormers.storm.network.RetrofitClient
-import com.stormers.storm.project.fragment.WaitingForStartingProjectFragment
-import com.stormers.storm.project.model.ResponseProjectInfoModel
+import com.stormers.storm.project.network.response.ResponseProjectInfoModel
 import com.stormers.storm.round.fragment.MemberWaitingFragment
 import kotlinx.android.synthetic.main.activity_round_setting.*
 import retrofit2.Call
@@ -19,7 +18,7 @@ import retrofit2.Response
 
 open class BaseProjectWaitingActivity : BaseActivity() {
 
-    private lateinit var retrofitClient: InterfaceProjectInfo
+    private lateinit var retrofitClient: RequestProject
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,7 +26,7 @@ open class BaseProjectWaitingActivity : BaseActivity() {
 
         goToFragment(MemberWaitingFragment::class.java, null)
 
-        retrofitClient = RetrofitClient.create(InterfaceProjectInfo::class.java)
+        retrofitClient = RetrofitClient.create(RequestProject::class.java)
 
         preference.getProjectIdx()?.let {
