+++ app/src/main/java/com/stormers/storm/project/network/response/ResponseProjectUserListModel.kt
@@ -1,4 +1,4 @@
-package com.stormers.storm.project.model
+package com.stormers.storm.project.network.response
 
 import com.stormers.storm.user.UserModel
 import java.util.function.BinaryOperator
