+++ app/src/main/java/com/stormers/storm/round/model/RoundDescriptionModel.kt
@@ -2,7 +2,6 @@ package com.stormers.storm.round.model
 
 import androidx.room.Ignore
 import com.google.gson.annotations.SerializedName
-import com.stormers.storm.round.network.RoundData
 
 class RoundDescriptionModel (
     //Todo: primitive 자료형으로 변경한 뒤, getter setter 등으로 데이터 가공할 수 있도록 변경
