--- app/src/main/java/com/stormers/storm/round/network/InterfaceRoundEnter.kt
@@ -1,13 +1,0 @@
-package com.stormers.storm.round.network
-
-import com.stormers.storm.network.BaseResponse
-import com.stormers.storm.network.SimpleResponse
-import com.stormers.storm.round.model.RoundEnterModel
-import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface InterfaceRoundEnter {
-    @POST("round/enter")
-    fun interfaceRoundEnter(@Body body :RoundEnterModel):Call<SimpleResponse>
-}
