+++ app/src/main/java/com/stormers/storm/ui/MainActivity.kt
@@ -14,12 +14,12 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.stormers.storm.R
 import com.stormers.storm.base.BaseActivity
-import com.stormers.storm.network.InterfaceJoinProjectUsingCode
-import com.stormers.storm.network.ResponseJoinProjectUsingCode
+import com.stormers.storm.project.network.RequestProject
+import com.stormers.storm.project.network.response.ResponseJoinProjectUsingCode
 import com.stormers.storm.network.RetrofitClient
 import com.stormers.storm.project.adapter.ParticipatedProjectListAdapter
 import com.stormers.storm.project.model.*
-import com.stormers.storm.project.network.ProjectInterface
+import com.stormers.storm.project.network.response.ResponseParticipatedProject
 import com.stormers.storm.util.MarginDecoration
 import kotlinx.android.synthetic.main.activity_main.*
 import retrofit2.Call
@@ -108,7 +108,7 @@ class MainActivity : BaseActivity() {
 
     private fun loadProjectsDatas() {
 
-        RetrofitClient.create(ProjectInterface::class.java).requestParticipatedProject(preference.getUserIdx()!!)
+        RetrofitClient.create(RequestProject::class.java).requestParticipatedProject(preference.getUserIdx()!!)
             .enqueue(object: Callback<ResponseParticipatedProject> {
                 override fun onFailure(call: Call<ResponseParticipatedProject>, t: Throwable) {
                     Log.d("requestParticipatedPj", "fail : ${t.message}")
@@ -148,7 +148,7 @@ class MainActivity : BaseActivity() {
 
             KeyEvent.KEYCODE_ENTER -> {
 
-                RetrofitClient.create(InterfaceJoinProjectUsingCode::class.java)
+                RetrofitClient.create(RequestProject::class.java)
                     .joinProjectUsingCode(JoinProjectUsingCodeModel(preference.getUserIdx()!!, edittext_input_participate_code.text.toString()))
                     .enqueue(object : Callback<ResponseJoinProjectUsingCode> {
