+++ app/src/main/java/com/stormers/storm/ui/ParticipatedProjectListActivity.kt
@@ -4,20 +4,16 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import com.stormers.storm.R
-import com.stormers.storm.project.model.RecentProjectsModel
 import com.stormers.storm.base.BaseActivity
 import com.stormers.storm.network.RetrofitClient
 import com.stormers.storm.project.adapter.ParticipatedProjectListAdapter
-import com.stormers.storm.project.model.ParticipatedProjectModel
-import com.stormers.storm.project.model.ResponseParticipatedProject
-import com.stormers.storm.project.network.ProjectInterface
+import com.stormers.storm.project.network.response.ResponseParticipatedProject
+import com.stormers.storm.project.network.RequestProject
 import com.stormers.storm.util.MarginDecoration
 import kotlinx.android.synthetic.main.activity_participated_project_list.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import java.util.*
-import java.util.Arrays.asList
 
 class ParticipatedProjectListActivity : BaseActivity() {
 
@@ -49,7 +45,7 @@ class ParticipatedProjectListActivity : BaseActivity() {
         //Todo: 로그인 이후 userId를 입력하여야 함
         preference.setUserId(1)
 
-        RetrofitClient.create(ProjectInterface::class.java).requestParticipatedProject(preference.getUserId()!!)
+        RetrofitClient.create(RequestProject::class.java).requestParticipatedProject(preference.getUserId()!!)
             .enqueue(object: Callback<ResponseParticipatedProject> {
                 override fun onFailure(call: Call<ResponseParticipatedProject>, t: Throwable) {
                     Log.d("requestParticipatedPj", "fail : ${t.message}")
